Host: winner messaging screen
The Host app now has a modal fullscreen dialog that shows the winners, and allows you to send a message (a random number) to one of them. The Player app will show this number, allowing us to confirm the winner (as access to the document will be protected through the UID).